moved registration tests out into separate files
